Selesai Modify UI Mood Journaling
