--- momease-front-end/lib/relaxation.dart
@@ -1,25 +1,0 @@
-import 'package:flutter/material.dart';
-
-class RelaxationPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Relaxation'),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(Icons.self_improvement, size: 100, color: Colors.green),
-            SizedBox(height: 20),
-            Text(
-              'Take a moment to relax!',
-              style: TextStyle(fontSize: 18),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
